refactor: move WithConfig to config.ts
