+++ src/api/gnssOffsetCorrector/index.ts
@@ -1,7 +1,8 @@
 import { Delta, Path, PathValue, SourceRef, Update } from '@signalk/server-api'
 import { IRouter } from 'express'
 import { get as _get } from 'lodash'
-import { SignalKMessageHub, WithConfig } from '../../app'
+import { SignalKMessageHub } from '../../app'
+import { WithConfig } from '../../config/config'
 import { ConfigApp } from '../../config/config'
 import { WithSecurityStrategy } from '../../security'
 import { correctPosition, AntennaOffset, Position } from './leverArm'
