+++ src/app.ts
@@ -2,7 +2,6 @@
 import { Delta, ServerAPI, SKVersion, FullSignalK } from '@signalk/server-api'
 import { EventEmitter } from 'node:events'
 
-import { Config, SettingsKeyUpdate, SettingsMutator } from './config/config'
 import DeltaCache from './deltacache'
 import { StalenessEnforcer } from './staleness'
 
@@ -34,13 +33,4 @@ export interface SignalKMessageHub extends EventEmitter {
     delta: Partial<Delta>,
     skVersion?: SKVersion
   ) => void
-}
-
-export interface WithConfig {
-  config: Config
-  /** Persist a scoped settings change, committing to `config.settings`
-   * only after the file write succeeds. See `applySettingsUpdate`. */
-  updateSettings: (
-    update: SettingsMutator | SettingsKeyUpdate[]
-  ) => Promise<void>
 }
