+++ src/index.ts
@@ -42,7 +42,8 @@ import _ from 'lodash'
 import path from 'path'
 import { startApis } from './api'
 import type { HistoryApiHttpRegistry } from './api/history'
-import { ServerApp, SignalKMessageHub, WithConfig } from './app'
+import { ServerApp, SignalKMessageHub } from './app'
+import { WithConfig } from './config/config'
 import { ConfigApp, load, sendBaseDeltas } from './config/config'
 import { createDebug } from './debug'
 import DeltaCache, { buildSrcToCanonicalMap } from './deltacache'
