+++ src/security.ts
@@ -31,7 +31,7 @@ import _ from 'lodash'
 import path from 'path'
 import { generate } from 'selfsigned'
 import { Mode } from 'stat-mode'
-import { WithConfig } from './app'
+import { WithConfig } from './config/config'
 import { createDebug } from './debug'
 import { LoginRateLimiter } from './login-rate-limiter'
 import dummysecurity from './dummysecurity'
